Evaluate "-3".
The diff is somewhat longish, so a commentary is in order.

It has become clear that the part of the evaluator that transforms
numerals to Integer numbers is the inverse of the part of the parser
that transforms numbers to numerals, so it would be desirable to check
these particular pieces of functionality against each other. At the same
time, the partiality in the current implementation has become obvious:
the logic of rendering numbers is enacted on a subset of L consisting of
Z, S and P only. So, I decided to extract these constructors to a
separate type, so that a partial function L -> Integer might become a
total function N -> Integer that is straightforward to check. The rest
of the commit is the checking machinery.